Shared folder renamed to ui
