+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../src/shared/globals.scss'
+import '../src/ui/globals.scss'
 import { AppProps } from 'next/app'
 
 function CustomApp({ Component, pageProps }: AppProps) {
