--- src/shared/body/Body.module.scss
@@ -1,8 +1,0 @@
-.body {
-    font-family: var(--fonts-body);
-    font-size: var(--font-sizes-medium, 1rem);
-    font-weight: var(--font-weights-normal, 300);
-    line-height: var(--line-heights-medium, 1.5);
-    color: var(--colors-gray-darker, #4F4F4F);
-    background-color: var(--colors-body-background, white);  
-}
