+++ src/ui/body/Body.module.scss
@@ -1,0 +1,8 @@
+.body {
+    font-family: var(--fonts-body);
+    font-size: var(--font-sizes-medium);
+    font-weight: var(--font-weights-normal);
+    line-height: var(--line-heights-medium);
+    color: var(--colors-gray-darker);
+    background-color: var(--colors-body-background);  
+}
