+++ src/ui/column/Column.module.scss
@@ -1,6 +1,6 @@
 .column {
-    --column-spacing-x: var(--space-medium, 1.5rem);
-    --column-spacing-y: var(--space-large, 3rem);
+    --column-spacing-x: var(--space-medium);
+    --column-spacing-y: var(--space-large);
 
     position: relative;
     max-width: 80ch;
