+++ src/ui/header/Header.module.scss
@@ -1,8 +1,8 @@
 .header {
     --header-z-index: 1030;
-    --header-spacing-y: var(--space-medium, 1.5rem);
-    --header-background-color: var(--colors-body-background, white);
-    --header-logo-color: var(--colors-gray-darkest, black);
+    --header-spacing-y: var(--space-medium);
+    --header-background-color: var(--colors-body-background);
+    --header-logo-color: var(--colors-gray-darkest);
 
     position: sticky;
     top: 0;
