+++ src/ui/timeline/Timeline.module.scss
@@ -1,7 +1,7 @@
 .timeline {
-    --timeline-spacing: var(--space-large, 3rem);
-    --timeline-stroke-color: var(--colors-gray-lighter, #bdbdbd);
-    --timeline-stroke-color-active: var(--colors-primary, #9B51E0);
+    --timeline-spacing: var(--space-large);
+    --timeline-stroke-color: var(--colors-gray-lighter);
+    --timeline-stroke-color-active: var(--colors-primary);
     --timeline-stroke-width: 2px;
     --timeline-dot-size: 14px;
